refactor: move CompGen to config and stop parsing config
Both changes requested in review.
The output generator for bash,fish,zsh is unchanged.

Also do minor whitespace and style adjustments as requested in review.
