+++ mkosi/completion.py
@@ -2,14 +2,59 @@
 
 import argparse
 import dataclasses
+import enum
 import io
 import shlex
 from collections.abc import Iterable, Mapping
 from pathlib import Path
 from typing import Optional, Union
 
-from mkosi.config import SETTINGS, SETTINGS_LOOKUP_BY_DEST, Args, CompGen, Verb, create_argument_parser
+from mkosi import config
 from mkosi.log import die
+from mkosi.util import StrEnum
+
+
+class CompGen(StrEnum):
+    default = enum.auto()
+    files   = enum.auto()
+    dirs    = enum.auto()
+
+    @staticmethod
+    def from_action(action: argparse.Action) -> "CompGen":
+        if isinstance(action.default, Path):
+            if action.default.is_dir():
+                return CompGen.dirs
+            else:
+                return CompGen.files
+        # TODO: the type of action.type is Union[Callable[[str], Any], FileType]
+        # the type of Path is type, but Path also works in this position,
+        # because the constructor is a callable from str -> Path
+        elif action.type is not None and (isinstance(action.type, type) and issubclass(action.type, Path)):  # type: ignore
+            if isinstance(action.default, Path) and action.default.is_dir():  # type: ignore
+                return CompGen.dirs
+            else:
+                return CompGen.files
+
+        return CompGen.default
+
+    def to_bash(self) -> str:
+        return f"_mkosi_compgen_{self}"
+
+    def to_fish(self) -> str:
+        if self == CompGen.files:
+            return "--force-files"
+        elif self == CompGen.dirs:
+            return "--force-files -a '(__fish_complete_directories)'"
+        else:
+            return "-f"
+
+    def to_zsh(self) -> str:
+        if self == CompGen.files:
+            return ":path:_files -/"
+        elif self == CompGen.dirs:
+            return ":directory:_files -f"
+        else:
+            return ""
 
 
 @dataclasses.dataclass(frozen=True)
@@ -23,7 +68,7 @@ class CompletionItem:
 
 
 def collect_completion_arguments() -> list[CompletionItem]:
-    parser = create_argument_parser()
+    parser = config.create_argument_parser()
 
     options = [
         CompletionItem(
@@ -37,7 +82,7 @@ def collect_completion_arguments() -> list[CompletionItem]:
         for action in parser._actions
         if (action.option_strings and
             action.help != argparse.SUPPRESS and
-            action.dest not in SETTINGS_LOOKUP_BY_DEST)
+            action.dest not in config.SETTINGS_LOOKUP_BY_DEST)
     ]
 
     options += [
@@ -47,9 +92,9 @@ def collect_completion_arguments() -> list[CompletionItem]:
             help=setting.help,
             nargs=setting.nargs or 1,
             choices=[str(c) for c in setting.choices] if setting.choices is not None else [],
-            compgen=setting.compgen,
+            compgen=CompGen.default,
         )
-        for setting in SETTINGS
+        for setting in config.SETTINGS
     ]
 
     return options
@@ -91,7 +136,7 @@ def finalize_completion_bash(options: list[CompletionItem], resources: Path) -> 
         c.write(compgen)
         c.write("\n\n")
 
-        c.write(to_bash_array("_mkosi_verbs", [str(v) for v in Verb]))
+        c.write(to_bash_array("_mkosi_verbs", [str(v) for v in config.Verb]))
         c.write("\n\n\n")
 
         c.write(completion.read_text())
@@ -105,7 +150,7 @@ def finalize_completion_fish(options: list[CompletionItem], resources: Path) -> 
         c.write("complete -c mkosi -f\n")
 
         c.write("complete -c mkosi -n '__fish_is_first_token' -a \"")
-        c.write(" ".join(str(v) for v in Verb))
+        c.write(" ".join(str(v) for v in config.Verb))
         c.write("\"\n")
 
         for option in options:
@@ -128,6 +173,7 @@ def finalize_completion_fish(options: list[CompletionItem], resources: Path) -> 
                 c.write(f"-d \"{help}\" ")
             c.write(option.compgen.to_fish())
             c.write("\n")
+
         return c.getvalue()
 
 
@@ -141,9 +187,8 @@ def finalize_completion_zsh(options: list[CompletionItem], resources: Path) -> s
         c.write("#compdef mkosi\n")
         c.write("# SPDX-License-Identifier: LGPL-2.1-or-later\n\n")
 
-        c.write(to_zsh_array("_mkosi_verbs", [str(v) for v in Verb]))
+        c.write(to_zsh_array("_mkosi_verbs", [str(v) for v in config.Verb]))
         c.write("\n\n")
-
 
         c.write(completion.read_text())
         c.write("\n")
@@ -179,7 +224,7 @@ def finalize_completion_zsh(options: list[CompletionItem], resources: Path) -> s
         return c.getvalue()
 
 
-def print_completion(args: Args, *, resources: Path) -> None:
+def print_completion(args: config.Args, *, resources: Path) -> None:
     if not args.cmdline:
         die(
             "No shell to generate completion script for specified",
@@ -187,15 +232,17 @@ def print_completion(args: Args, *, resources: Path) -> None:
         )
 
     shell = args.cmdline[0]
-    completion_args = collect_completion_arguments()
     if shell == "bash":
-        print(finalize_completion_bash(completion_args, resources))
+        func = finalize_completion_bash
     elif shell == "fish":
-        print(finalize_completion_fish(completion_args, resources))
+        func = finalize_completion_fish
     elif shell == "zsh":
-        print(finalize_completion_zsh(completion_args, resources))
+        func = finalize_completion_zsh
     else:
         die(
             f"{shell!r} is not supported for completion scripts.",
             hint="Please specify either one of: bash, fish, zsh"
         )
+
+    completion_args = collect_completion_arguments()
+    print(func(completion_args, resources))
